+++ backend/connector-integration/src/connectors.rs
@@ -101,8 +101,14 @@ pub use self::payload::Payload;
 pub mod fiservemea;
 pub use self::fiservemea::Fiservemea;
 
+pub mod paysafe;
+pub use self::paysafe::Paysafe;
+
 pub mod datatrans;
 pub use self::datatrans::Datatrans;
+
+pub mod bluesnap;
+pub use self::bluesnap::Bluesnap;
 
 pub mod authipay;
 pub use self::authipay::Authipay;