+++ backend/domain_types/src/payment_address.rs
@@ -2,7 +2,9 @@ use common_enums::ProductType;
 use common_utils::{ext_traits::ConfigExt, Email, MinorUnit};
 use hyperswitch_masking::{PeekInterface, Secret, SerializableSecret};
 
-use crate::utils::{convert_us_state_to_code, missing_field_err, Error};
+use crate::utils::{
+    convert_canada_state_to_code, convert_us_state_to_code, missing_field_err, Error,
+};
 
 #[derive(Clone, Default, Debug)]
 pub struct PaymentAddress {
@@ -318,6 +320,9 @@ impl AddressDetails {
         match country {
             common_enums::CountryAlpha2::US => Ok(Secret::new(
                 convert_us_state_to_code(&state.peek().to_string()).to_string(),
+            )),
+            common_enums::CountryAlpha2::CA => Ok(Secret::new(
+                convert_canada_state_to_code(&state.peek().to_string()).to_string(),
             )),
             _ => Ok(state.clone()),
         }